requirements.txt
absl-py==2.3.1
annotated-types==0.7.0
anyio==4.12.0
asgiref==3.11.0
astunparse==1.6.3
attrs==25.4.0
autobahn==24.4.2
Automat==22.10.0
certifi==2025.11.12
cffi==2.0.0
channels==4.3.2
charset-normalizer==3.4.4
cloudpickle==3.1.2
constantly==23.10.4
cryptography==46.0.3
daphne==4.2.1
distro==1.9.0
dj-database-url==3.0.1
Django==5.2.8
django-environ==0.12.0
exceptiongroup==1.3.1
flatbuffers==25.9.23
gast==0.7.0
google-ai-generativelanguage==0.6.16
google-api-core==2.23.0
google-api-python-client==2.161.0
google-generativeai==0.8.5
google-pasta==0.2.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
grpcio==1.76.0
grpcio-status==1.62.3
gunicorn==23.0.0
h11==0.16.0
h5py==3.15.1
httpcore==1.0.9
httpx==0.28.1
httplib2==0.22.0
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
joblib==1.5.2
keras==3.12.0
# LangChain + Gemini integration
langchain>=0.3.0
langchain-core>=0.3.0
langchain-google-genai>=2.0.0
libclang==18.1.1
llvmlite==0.45.1
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ml_dtypes==0.5.4
namex==0.1.0
numba==0.62.1
numpy==2.2.6
opt_einsum==3.4.0
optree==0.18.0
packaging==25.0
pandas==2.3.3
pillow==12.0.0
proto-plus==1.26.0
protobuf>=3.20.3,<5.0.0
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pygments==2.19.2
pyOpenSSL==25.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
# YAML config support
PyYAML>=6.0
requests==2.32.5
rich==14.2.0
rsa==4.9
scikit-learn==1.7.2
scipy==1.15.3
service-identity==24.2.0
shap==0.49.1
six==1.17.0
slicer==0.0.8
sniffio==1.3.1
sqlparse==0.5.5
ta==0.11.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.2.0
threadpoolctl==3.6.0
tomli==2.3.0
tqdm==4.67.1
Twisted==25.5.0
txaio==25.9.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate>=4.1.1
urllib3==2.5.0
vaderSentiment==3.3.2
Werkzeug==3.1.4
whitenoise==6.11.0
wrapt==2.0.1
# XGBoost for hybrid model volatile-regime fallback
xgboost>=2.0.0
zope.interface==8.1.1
